+++ tests/test_void_and_cluster.py
@@ -2,9 +2,11 @@ import pytest
 
 import numpy as np
 
-from vc_sample.density_estimation import KernelDensityEstimator
+from vc_sample.density_estimation import KernelDensityEstimator, UMAPDensityEstimator
 from vc_sample.kernels import GaussianKernel
 from vc_sample.void_and_cluster import VoidAndCluster
+
+from sklearn.datasets import make_circles
 
 
 @pytest.mark.parametrize("sample_size", [10, 20, 50, 100])
@@ -26,3 +28,16 @@ def test_sample_size(sample_size):
 
     assert len(samples) == len(ordering) == sample_size
     assert np.unique(samples, axis=0).shape[0] == sample_size
+
+
+def test_umap_circle():
+    circles, circles_class = make_circles(n_samples=100, random_state=42)
+    estimator = UMAPDensityEstimator(X=circles, n_neighbors=10)
+
+    vc = VoidAndCluster(estimator, circles.shape[0], num_initial_samples=10)
+
+    sample_indices = vc.sample(size=20)
+    assert len(sample_indices) == 20
+
+    circles_class_samples = circles_class[sample_indices]
+    assert np.unique(circles_class_samples).shape[0] == 2