+++ CHANGELOG.rst
@@ -8,4 +8,4 @@ Version 0.01
 Implementation of basic void-and-cluster algorithm:
 
 - Simple, unoptimized implementation using NumPy
-- Only kernel density estimation is supported (Specifying the kernel size is quite cumbersome)
+- Implementation of kernel and UMAP density estimator